// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPPORTER: 'SUPPORTER',
  MODERATOR: 'MODERATOR',
  ADMIN: 'ADMIN',
  CLUB_REP: 'CLUB_REP'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const LoyaltyTier = {
  NEW_FAN: 'NEW_FAN',
  SUPPORTER: 'SUPPORTER',
  CORE_SUPPORTER: 'CORE_SUPPORTER',
  ULTRAS: 'ULTRAS'
} as const

export type LoyaltyTier = (typeof LoyaltyTier)[keyof typeof LoyaltyTier]


export const FixtureStatus = {
  SCHEDULED: 'SCHEDULED',
  LIVE: 'LIVE',
  COMPLETED: 'COMPLETED',
  POSTPONED: 'POSTPONED',
  CANCELLED: 'CANCELLED'
} as const

export type FixtureStatus = (typeof FixtureStatus)[keyof typeof FixtureStatus]